change(common): move path::MapKey out of bpf module
